Updated Register, CenterScreen, FunctionEvents Class
Changed:
- Moved FunctionEvents Class into a seperate file since its over 200 lines of code (One has 100 lines)
- Moved Registration Screen into a separate form

Updated:
- Fixed Centering of Forms. Now forms that should center the parent form will work
