+++ GameLauncher/App/Classes/LauncherCore/Global/FunctionStatus.cs
@@ -1,7 +1,6 @@
 using GameLauncher.App.Classes.LauncherCore.FileReadWrite;
 using GameLauncher.App.Classes.LauncherCore.Lists;
 using GameLauncher.App.Classes.LauncherCore.Lists.JSON;
-using GameLauncher.App.Classes.LauncherCore.Visuals;
 using GameLauncher.App.Classes.Logger;
 using GameLauncher.App.Classes.SystemPlatform.Linux;
 using GameLauncher.App.Classes.SystemPlatform.Windows;
@@ -14,7 +13,6 @@ using System.Globalization;
 using System.IO;
 using System.Management.Automation;
 using System.Net;
-using System.Text;
 using System.Windows.Forms;
 
 namespace GameLauncher.App.Classes.LauncherCore.Global
@@ -22,6 +20,9 @@ namespace GameLauncher.App.Classes.LauncherCore.Global
     /* This is Used to Cache Responses From the Launcher */
     class InformationCache
     {
+        /* Parent Screen Cords */
+        public static Point ParentScreenLocation;
+
         /* ServerList Load Checks */
         public static string ServerListStatus = "Unknown";
 
@@ -36,6 +37,9 @@ namespace GameLauncher.App.Classes.LauncherCore.Global
 
         /* System OS Name */
         public static string OSName;
+
+        /* Selected Server Auth Support */
+        public static bool ModernAuthSupport = false;
 
         /* Selected Server Category */
         public static string SelectedServerCategory;
@@ -53,6 +57,9 @@ namespace GameLauncher.App.Classes.LauncherCore.Global
     /* This is Used to call Certain Functions (Such as Completion Status or Function Callbacks) */
     class FunctionStatus
     {
+        /* Allows Registration Button to be Enabled/Disabled */
+        public static bool AllowRegistration;
+
         /* Verify Hash Status */
         public static bool IsVerifyHashDisabled = false;
 
@@ -87,6 +94,12 @@ namespace GameLauncher.App.Classes.LauncherCore.Global
             form.StartPosition = FormStartPosition.Manual;
             form.Top = (Screen.PrimaryScreen.Bounds.Height - form.Height) / 2;
             form.Left = (Screen.PrimaryScreen.Bounds.Width - form.Width) / 2;
+        }
+
+        public static void CenterParent(Form form)
+        {
+            form.StartPosition = FormStartPosition.Manual;
+            form.Location = InformationCache.ParentScreenLocation;
         }
 
         /* Check if Folder Location is Acceptable and Returns a Value
@@ -176,7 +189,7 @@ namespace GameLauncher.App.Classes.LauncherCore.Global
 
                 if (fbd.ShowDialog() == CommonFileDialogResult.Ok)
                 {
-                    if (!FunctionStatus.HasWriteAccessToFolder(fbd.FileName))
+                    if (!HasWriteAccessToFolder(fbd.FileName))
                     {
                         Log.Error("LAUNCHER: Not enough permissions. Exiting.");
                         MessageBox.Show(null, "You don't have enough permission to select this path as installation folder. Please select another directory.", "GameLauncher", MessageBoxButtons.OK, MessageBoxIcon.Information);
@@ -217,7 +230,7 @@ namespace GameLauncher.App.Classes.LauncherCore.Global
 
             if (!DetectLinux.LinuxDetected())
             {
-                switch (FunctionStatus.CheckFolder(FileSettingsSave.GameInstallation))
+                switch (CheckFolder(FileSettingsSave.GameInstallation))
                 {
                     case FolderType.IsSameAsLauncherFolder:
                         Directory.CreateDirectory("Game Files");
@@ -292,152 +305,5 @@ namespace GameLauncher.App.Classes.LauncherCore.Global
 
         /* Moved "runAsAdmin" Code to Gist */
         /* https://gist.githubusercontent.com/DavidCarbon/97494268b0175a81a5f89a5e5aebce38/raw/eec2f9f80aa4b350ab98d32383e1ee1f2e1c26fd/Self.cs */
-    }
-
-    /* This is Used to Visual Events (Draw or Click Events) */
-    class FunctionEvents
-    {
-        /* ServerList Load Checks */
-        public static void SelectServerBtn_Click(object sender, EventArgs e)
-        {
-            new SelectServer().ShowDialog();
-        }
-
-        public static void DiscordInviteLink_LinkClicked(object sender, LinkLabelLinkClickedEventArgs e)
-        {
-            if (!string.IsNullOrEmpty(InformationCache.SelectedServerJSON.discordUrl))
-                Process.Start(InformationCache.SelectedServerJSON.discordUrl);
-        }
-
-        public static void HomePageLink_LinkClicked(object sender, LinkLabelLinkClickedEventArgs e)
-        {
-            if (!string.IsNullOrEmpty(InformationCache.SelectedServerJSON.homePageUrl))
-                Process.Start(InformationCache.SelectedServerJSON.homePageUrl);
-        }
-
-        public static void FacebookGroupLink_LinkClicked(object sender, LinkLabelLinkClickedEventArgs e)
-        {
-            if (!string.IsNullOrEmpty(InformationCache.SelectedServerJSON.facebookUrl))
-                Process.Start(InformationCache.SelectedServerJSON.facebookUrl);
-        }
-
-        public static void TwitterAccountLink_LinkClicked(object sender, LinkLabelLinkClickedEventArgs e)
-        {
-            if (!string.IsNullOrEmpty(InformationCache.SelectedServerJSON.facebookUrl))
-                Process.Start(InformationCache.SelectedServerJSON.twitterUrl);
-        }
-
-        public static void ForgotPassword_LinkClicked(object sender, LinkLabelLinkClickedEventArgs e)
-        {
-            if (!string.IsNullOrEmpty(InformationCache.SelectedServerJSON.webRecoveryUrl))
-            {
-                Process.Start(InformationCache.SelectedServerJSON.webRecoveryUrl);
-                MessageBox.Show(null, "A browser window has been opened to complete password recovery on " + 
-                    InformationCache.SelectedServerJSON.serverName, "GameLauncher", MessageBoxButtons.OK, MessageBoxIcon.Information);
-                return;
-            }
-            else
-            {
-                string send = Prompt.ShowDialog("Please specify your email address.", "GameLauncher");
-
-                if (send != String.Empty)
-                {
-                    String responseString;
-                    try
-                    {
-                        Uri resetPasswordUrl = new Uri(InformationCache.SelectedServerData.IpAddress + "/RecoveryPassword/forgotPassword");
-
-                        var request = (HttpWebRequest)System.Net.WebRequest.Create(resetPasswordUrl);
-                        var postData = "email=" + send;
-                        var data = Encoding.ASCII.GetBytes(postData);
-                        request.Method = "POST";
-                        request.ContentType = "application/x-www-form-urlencoded";
-                        request.ContentLength = data.Length;
-
-                        using (var stream = request.GetRequestStream())
-                        {
-                            stream.Write(data, 0, data.Length);
-                        }
-
-                        var response = (HttpWebResponse)request.GetResponse();
-                        responseString = new StreamReader(response.GetResponseStream()).ReadToEnd();
-                    }
-                    catch
-                    {
-                        responseString = "Failed to send email!";
-                    }
-
-                    MessageBox.Show(null, responseString, "GameLauncher", MessageBoxButtons.OK, MessageBoxIcon.Information);
-                }
-            }
-        }
-
-        public static void ComboBox1_DrawItem(object sender, DrawItemEventArgs e)
-        {
-            var font = (sender as ComboBox).Font;
-            Brush backgroundColor;
-            Brush textColor;
-
-            var serverListText = "";
-            int onlineStatus = 2; /* 0 = offline | 1 = online | 2 = checking */
-
-            if (sender is ComboBox cb)
-            {
-                if (cb.Items[e.Index] is ServerList si)
-                {
-                    serverListText = si.Name;
-                    onlineStatus = InformationCache.ServerStatusBook.ContainsKey(si.Id) ? InformationCache.ServerStatusBook[si.Id] : 2;
-                }
-            }
-
-            if (serverListText.StartsWith("<GROUP>"))
-            {
-                font = new Font(font, FontStyle.Bold);
-                e.Graphics.FillRectangle(Brushes.White, e.Bounds);
-                e.Graphics.DrawString(serverListText.Replace("<GROUP>", string.Empty), font, Brushes.Black, e.Bounds);
-            }
-            else
-            {
-                font = new Font(font, FontStyle.Regular);
-                if ((e.State & DrawItemState.Selected) == DrawItemState.Selected && e.State != DrawItemState.ComboBoxEdit)
-                {
-                    backgroundColor = SystemBrushes.Highlight;
-                    textColor = SystemBrushes.HighlightText;
-                }
-                else
-                {
-                    if (onlineStatus == 2)
-                    {
-                        /* CHECKING */
-                        backgroundColor = Brushes.Khaki;
-                    }
-                    else if (onlineStatus == 1)
-                    {
-                        /* ONLINE */
-                        backgroundColor = Brushes.PaleGreen;
-                    }
-                    else
-                    {
-                        /* OFFLINE */
-                        backgroundColor = Brushes.LightCoral;
-                    }
-
-                    textColor = Brushes.Black;
-                }
-
-                e.Graphics.FillRectangle(backgroundColor, e.Bounds);
-                e.Graphics.DrawString("    " + serverListText, font, textColor, e.Bounds);
-            }
-        }
-
-        public static void AddServer_Click(object sender, EventArgs e)
-        {
-            new AddServer().Show();
-        }
-
-        public static void AboutButton_Click(object sender, EventArgs e)
-        {
-            new About().ShowDialog();
-        }
     }
 }
