+++ GameLauncher/GameLauncher.csproj
@@ -165,6 +165,7 @@
     <Compile Include="App\Classes\LauncherCore\FileReadWrite\FileAccountSave.cs" />
     <Compile Include="App\Classes\LauncherCore\FileReadWrite\FileSettingsSave.cs" />
     <Compile Include="App\Classes\LauncherCore\Global\EnumTypes.cs" />
+    <Compile Include="App\Classes\LauncherCore\Global\FunctionEvents.cs" />
     <Compile Include="App\Classes\LauncherCore\Global\FunctionStatus.cs" />
     <Compile Include="App\Classes\LauncherCore\Global\URLs.cs" />
     <Compile Include="App\Classes\LauncherCore\Hash\SHATwoFiveSix.cs" />
@@ -227,6 +228,12 @@
     <Compile Include="App\Classes\LauncherCore\RPC\DiscordGamePresence.cs" />
     <Compile Include="App\Classes\LauncherCore\RPC\EventList.cs" />
     <Compile Include="App\Classes\SystemPlatform\Windows\CertificateStore.cs" />
+    <Compile Include="App\RegisterScreen.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="App\RegisterScreen.Designer.cs">
+      <DependentUpon>RegisterScreen.cs</DependentUpon>
+    </Compile>
     <Compile Include="App\SettingsScreen.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -313,6 +320,9 @@
       <DependentUpon>About.cs</DependentUpon>
       <SubType>Designer</SubType>
     </EmbeddedResource>
+    <EmbeddedResource Include="App\RegisterScreen.resx">
+      <DependentUpon>RegisterScreen.cs</DependentUpon>
+    </EmbeddedResource>
     <EmbeddedResource Include="App\SettingsScreen.resx">
       <DependentUpon>SettingsScreen.cs</DependentUpon>
       <SubType>Designer</SubType>
@@ -389,6 +399,7 @@
     <None Include="Resources\icon_shield_warning.png" />
     <None Include="Resources\logo_main.png" />
     <None Include="Resources\logo_splash.png" />
+    <None Include="Resources\background_register.png" />
     <Content Include="SBRW.ico" />
     <None Include="Resources\progress_warning.png" />
     <None Include="Resources\progress_success.png" />
